Converted float values to Decimals

+++ coinbase_advanced_trader/models/past_order.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
 
 @dataclass
 class PastOrder:
@@ -7,7 +8,7 @@ class PastOrder:
     timestamp: datetime
     action: str
     order_id: str
-    nearest_risk: float
-    value: float
-    balance: float
+    nearest_risk: Decimal
+    value: Decimal
+    balance: Decimal
     status: str
